backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
mysqlclient>=2.2,<3.0
python-dotenv>=1.0,<2.0
xhtml2pdf>=0.2.16,<0.3.0
openpyxl>=3.1,<4.0
